requirements.txt
argparse
django
pillow
image
numpy
PyQt5